more tv handling
